requirements.txt
django
django-extensions
drf-yasg
djangorestframework
ipython
djangorestframework-simplejwt